src/main/webapp/app/entities/roditelj/route/roditelj-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IRoditelj, Roditelj } from '../roditelj.model';
import { RoditeljService } from '../service/roditelj.service';

import { RoditeljRoutingResolveService } from './roditelj-routing-resolve.service';

describe('Roditelj routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: RoditeljRoutingResolveService;
  let service: RoditeljService;
  let resultRoditelj: IRoditelj | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(RoditeljRoutingResolveService);
    service = TestBed.inject(RoditeljService);
    resultRoditelj = undefined;
  });

  describe('resolve', () => {
    it('should return IRoditelj returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultRoditelj = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultRoditelj).toEqual({ id: 123 });
    });

    it('should return new IRoditelj if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultRoditelj = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultRoditelj).toEqual(new Roditelj());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as Roditelj })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultRoditelj = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultRoditelj).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
